+++ src/edu/uclouvain/core/nodus/utils/LinkedProperties.java
@@ -1,3 +1,0 @@
-package edu.uclouvain.core.nodus.utils;
-
-public class LinkedProperties {}